chore(room_reservation): change workflow trigger from command to text (#97)
By switching to data.text, we avoid the need to register specific
commands, allowing users the flexibility to modify or add their own
commands. While this approach doesn’t offer all the benefits of a
traditional slash command, it provides greater adaptability. To handle
this, a utility function was added to parse the text, as the email
format returned by Slack was not directly usable.

Additionally, the README and docstrings were updated to reflect these
changes.