fix(samples): activate `start_github_action` and update `repo` to `repository` per eng-1631 (#117)
**Issues Addressed:**
- The `start_github_action` trigger function was previously commented
out to encourage users to configure the workflow file before running the
sample. This was relevant before cloud, where the `.yaml` file was
easily accessible, which is no longer the case. The solution is to keep
the trigger function active from the start. I updated the docstring to
include clear instructions reflecting this change.
- The push event was not received from the app because push events were
not activated in the app configuration.
- Per ENG-1631, I updated instances of `repo` to `repository` to align
with how GitHub structures JSON event data.