feat: port github_copilot_seats to python (#104)
`_add_seat`, `_remove_seat`, and `_get_all_seats()` utilize the
`Requester` object to make requests to APIs that PyGithub doesn't
currently support. To facilitate this, I updated the `pyproject.toml` to
`"PyGithub @ git+https://github.com/PyGithub/PyGithub.git"` in a
separate [PR](https://github.com/autokitteh/autokitteh/pull/802),
ensuring we get the latest version. This change is necessary because the
`Requester` object was recently added, and the last release was in
August.

I mention this here because both PRs will need to be merged
simultaneously.

---------

Co-authored-by: Itay Donanhirsh <itay@autokitteh.com>
Co-authored-by: Daniel Abraham <daniel@autokitteh.com>